convert actionMenu.js to typescript
