+++ ui/types/lichess-jquery.d.ts
@@ -1,0 +1,8 @@
+interface JQueryStatic {
+  modal: LichessModal;
+}
+
+interface LichessModal {
+  (html: string | JQuery): JQuery;
+  close(): void;
+}
